refactor(parser): split word helpers into subdomains
