+++ crates/shuck-parser/src/parser/mod.rs
@@ -23,8 +23,7 @@ mod result;
 mod source_tree;
 mod syntax_facts;
 mod token_stream;
-mod word_builders;
-mod word_parts;
+mod word;
 mod word_tokens;
 mod words;
 mod zsh_features;
